01-discovery/automation/requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.5
openai>=1.12.0
azure-identity>=1.15.0
pyyaml>=6.0
jsonschema>=4.20.0
markdown>=3.5
python-dotenv>=1.0.0
rich>=13.7.0

# Testing and coverage
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
coverage>=7.3.0
